control de video reutilizable
